Tweaks

+++ Rakefile
@@ -5,13 +5,13 @@ require './lib/indent'
 task default: [:usage] do
 end
 
-task js: [:clean_output, :convert] do
-  dst = '/output/icons.js'
+task js: [:clean_tmp, :convert] do
+  dst = '/app/output/icons.js'
   svgs = {}
-  Dir['/output/**/*.svg'].each do |file|
+  Dir['/app/tmp/**/*.svg'].each do |file|
     next if file =~ /colorMap\.svg/
 
-    name = file.gsub(/^output\//, '')
+    name = file.gsub(/^\/app\/tmp\//, '')
                .gsub(/\.svg$/, '')
     content = File.read(file)
                   .delete("\n") # Remove newlines
@@ -32,14 +32,14 @@ task js: [:clean_output, :convert] do
   puts "\nWritten to #{dst}"
 end
 
-task json: [:clean_output, :convert] do
-  dst = '/output/icons.json'
+task json: [:clean_tmp, :convert] do
+  dst = '/app/output/icons.json'
   svgs = {}
   svgs = {}
-  Dir['/output/**/*.svg'].each do |file|
+  Dir['/app/tmp/**/*.svg'].each do |file|
     next if file =~ /colorMap\.svg/
 
-    name = file.gsub(/^output\//, '')
+    name = file.gsub(/^\/app\/tmp\//, '')
                .gsub(/\.svg$/, '')
     content = File.read(file)
                   .delete("\n") # Remove newlines
@@ -59,12 +59,12 @@ task json: [:clean_output, :convert] do
   puts "\nWritten to #{dst}"
 end
 
-task preview: [:clean_output, :convert] do
+task preview: [:clean_tmp, :convert] do
   puts 'Create preview:'
-  dst = '/output/preview.html'
-  color_map = ColorMap.new('/input/colorMap.svg')
+  dst = '/app/output/preview.html'
+  color_map = ColorMap.new('/app/input/colorMap.svg')
   icons = ''
-  Dir['/output/**/*.svg'].each do |file|
+  Dir['/app/tmp/**/*.svg'].each do |file|
     next if file =~ /colorMap\.svg/
 
     content = File.read(file)
@@ -87,13 +87,14 @@ task preview: [:clean_output, :convert] do
   icons_indent = template.match(/^( +)<!-- ICONS -->/)[1].length
   template.gsub!(/^ +<!-- ICONS -->/, icons.indent(icons_indent))
   File.open(dst, 'w') { |f| f.write(template) }
-  # `open #{dst}`
+
+  puts '-> Done'
 end
 
 task :convert do
   puts 'Convert icons:'
-  color_map = ColorMap.new('/input/colorMap.svg')
-  Dir['/input/**/*.svg'].each do |file|
+  color_map = ColorMap.new('/app/input/colorMap.svg')
+  Dir['/app/input/**/*.svg'].each do |file|
     next if file =~ /colorMap\.svg/
 
     Convert.convert(file, color_map)
@@ -104,15 +105,15 @@ task clean: [:clean_input, :clean_output] do
 end
 
 task :clean_input do
-  Dir.chdir('/input')
+  Dir.chdir('/app/input')
 
   # Remove folders
-  Dir['/input/*/'].each do |dir|
+  Dir['/app/input/*/'].each do |dir|
     FileUtils.rm_rf(dir)
   end
 
   # Remove files
-  Dir['/input/**/*.*'].each do |file|
+  Dir['/app/input/**/*.*'].each do |file|
     File.delete(file)
   end
 
@@ -120,15 +121,35 @@ task :clean_input do
 end
 
 task :clean_output do
-  Dir.chdir('/output')
+  Dir.chdir('/app/output')
 
   # Remove folders
-  Dir['/output/*/'].each do |dir|
+  Dir['/app/output/*/'].each do |dir|
+    next if dir =~ /output\/input/
+
     FileUtils.rm_rf(dir)
   end
 
   # Remove files
-  Dir['/output/**/*.*'].each do |file|
+  Dir['/app/output/**/*.*'].each do |file|
+    next if file =~ /output\/input/
+
+    File.delete(file)
+  end
+
+  Dir.chdir('..')
+end
+
+task :clean_tmp do
+  Dir.chdir('/app/tmp')
+
+  # Remove folders
+  Dir['/app/tmp/*/'].each do |dir|
+    FileUtils.rm_rf(dir)
+  end
+
+  # Remove files
+  Dir['/app/tmp/**/*.*'].each do |file|
     File.delete(file)
   end
 
@@ -136,17 +157,13 @@ task :clean_output do
 end
 
 task :usage do
-  image_name = 'madebytight/svg-icons'
-  cmd = [
-    'docker run --rm',
-    '-v $(pwd)/input:/input',
-    '-v $(pwd)/output:/output',
-    image_name
-  ].join(' ')
+  puts <<~USAGE
+    Usage:
+      docker run --rm -v $(pwd)/input:/app/input -v $(pwd)/output:/app/output madebytight/svg-icons COMMAND
 
-
-  puts "\n\nUsage:\n"
-  puts "#{cmd} rake json     -- Export to JSON"
-  puts "#{cmd} rake js       -- Export to JavaScript"
-  puts "#{cmd} rake preview  -- Create preview in HTML"
+    Commands:
+      json      Export to JSON
+      js        Export to JavaScript
+      preview   Create preview in HTML
+  USAGE
 end
